refactor(ui): extract project settings panels
