+++ docs-ai/skills/ui/SKILL.md
@@ -174,8 +174,9 @@ Each section has exactly one job: orient, inspect, compare, edit, or confirm. If
 - Projects UI keeps Project Assistant orchestration in
   `useProjectAssistantController`; do not put propose/apply/bootstrap request
   state back into `ProjectsView`. The workspace rendering branches live behind
-  `ProjectWorkspaceView`, while project state loading and mutations stay in the
-  parent page.
+  `ProjectWorkspaceView`, and project defaults/root editing lives in
+  `ProjectSettingsPanel` plus `CreateProjectWorktreeModal`, while project state
+  loading and mutations stay in the parent page.
 - External Agent readiness belongs in Connections and in the picker
   diagnostics: distinguish missing binaries, auth/billing problems, unsupported
   versions, and managed-launcher issues without sending users to raw logs first.
